fix: extract get_client to a leaf client.py to break the import cycle
The @durable_task adapter added a workflows→task→dbos_app→workflows import cycle
(CodeQL py/cyclic-import). Move get_client + system_database_url into a new leaf
client.py (imports only constants); task.py/workflows.py enqueue via .client,
dbos_app re-exports get_client for back-compat and keeps configure_dbos. This
also breaks the pre-existing workflows↔dbos_app cycle. Registration timing +
behaviour unchanged (configure_dbos builds + registers verified; 79 tests green).
